Anomaly compt fix
Now the caverns should be fully compatible with the anomaly. Fleshbeast holes, noctolites and those things from the final quest should spawn normally (with the caveat that they can spawn in the player's base). Except for obelisks, since they fall from the sky, and there is no sky in the caverns :)